+++ packages/freezed/test/map_test.dart
@@ -28,43 +28,44 @@ void main() {
         'a second',
       );
     });
+
     group('default ctor', () {
       test("assert callbacks can't be null", () {
-        final value = Complex('a');
+        final value = SwitchTest('a');
 
         expect(
           () => value.map(
             (a) {},
             first: null,
-            second: (Complex2 value) {},
+            second: (SwitchTest2 value) {},
           ),
           throwsA(isA<AssertionError>()),
         );
         expect(
           () => value.map(
             null,
-            first: (Complex1 value) {},
-            second: (Complex2 value) {},
+            first: (SwitchTest1 value) {},
+            second: (SwitchTest2 value) {},
           ),
           throwsA(isA<AssertionError>()),
         );
         expect(
           () => value.map(
             (a) {},
-            first: (Complex1 value) {},
+            first: (SwitchTest1 value) {},
             second: null,
           ),
           throwsA(isA<AssertionError>()),
         );
       });
       test('calls default callback', () {
-        final value = Complex('a');
+        final value = SwitchTest('a');
 
         expect(
           value.map(
-            (Complex0 value) => '${value.a} 42',
-            first: (Complex1 value) => throw Error(),
-            second: (Complex2 value) => throw Error(),
+            (SwitchTest0 value) => '${value.a} 42',
+            first: (SwitchTest1 value) => throw Error(),
+            second: (SwitchTest2 value) => throw Error(),
           ),
           'a 42',
         );
@@ -72,41 +73,41 @@ void main() {
     });
     group('first ctor', () {
       test("assert callbacks can't be null", () {
-        final value = Complex.first('a');
+        final value = SwitchTest.first('a');
 
         expect(
           () => value.map(
             (a) {},
             first: null,
-            second: (Complex2 value) {},
+            second: (SwitchTest2 value) {},
           ),
           throwsA(isA<AssertionError>()),
         );
         expect(
           () => value.map(
             null,
-            first: (Complex1 value) {},
-            second: (Complex2 value) {},
+            first: (SwitchTest1 value) {},
+            second: (SwitchTest2 value) {},
           ),
           throwsA(isA<AssertionError>()),
         );
         expect(
           () => value.map(
             (a) {},
-            first: (Complex1 value) {},
+            first: (SwitchTest1 value) {},
             second: null,
           ),
           throwsA(isA<AssertionError>()),
         );
       });
       test('calls first callback', () {
-        final value = Complex.first('a', b: false, d: .42);
+        final value = SwitchTest.first('a', b: false, d: .42);
 
         expect(
           value.map(
-            (Complex0 a) => throw Error(),
-            first: (Complex1 value) => '${value.a} ${value.b} ${value.d}',
-            second: (Complex2 value) => throw Error(),
+            (SwitchTest0 a) => throw Error(),
+            first: (SwitchTest1 value) => '${value.a} ${value.b} ${value.d}',
+            second: (SwitchTest2 value) => throw Error(),
           ),
           'a false 0.42',
         );
@@ -114,41 +115,41 @@ void main() {
     });
     group('second ctor', () {
       test("assert callbacks can't be null", () {
-        final value = Complex.second('a');
+        final value = SwitchTest.second('a');
 
         expect(
           () => value.map(
             (a) {},
             first: null,
-            second: (Complex2 value) {},
+            second: (SwitchTest2 value) {},
           ),
           throwsA(isA<AssertionError>()),
         );
         expect(
           () => value.map(
             null,
-            first: (Complex1 value) {},
-            second: (Complex2 value) {},
+            first: (SwitchTest1 value) {},
+            second: (SwitchTest2 value) {},
           ),
           throwsA(isA<AssertionError>()),
         );
         expect(
           () => value.map(
             (a) {},
-            first: (Complex1 value) {},
+            first: (SwitchTest1 value) {},
             second: null,
           ),
           throwsA(isA<AssertionError>()),
         );
       });
       test('calls second callback', () {
-        final value = Complex.second('a', 21, .42);
+        final value = SwitchTest.second('a', 21, .42);
 
         expect(
           value.map(
-            (Complex0 a) => throw Error(),
-            first: (Complex1 value) => throw Error(),
-            second: (Complex2 value) => '${value.a} ${value.c} ${value.d}',
+            (SwitchTest0 a) => throw Error(),
+            first: (SwitchTest1 value) => throw Error(),
+            second: (SwitchTest2 value) => '${value.a} ${value.c} ${value.d}',
           ),
           'a 21 0.42',
         );
@@ -163,7 +164,7 @@ library main;
 import 'multiple_constructors.dart';
 
 void main() {
-  final value = Complex.first('a', b: false, d: .42);
+  final value = SwitchTest.first('a', b: false, d: .42);
 
   value.map(
     (String a) => 42,
@@ -183,7 +184,7 @@ void main() {
 
   group('maybeWhen', () {
     test('returns callback result if has callback', () {
-      var value = Complex('a');
+      var value = SwitchTest('a');
 
       expect(
         value.maybeMap((value) => '${value.a} default',
@@ -191,23 +192,23 @@ void main() {
         'a default',
       );
 
-      value = Complex.first('a', b: false, d: .42);
+      value = SwitchTest.first('a', b: false, d: .42);
 
       expect(
         value.maybeMap(
           null,
-          first: (Complex1 value) => '${value.a} ${value.b} ${value.d}',
+          first: (SwitchTest1 value) => '${value.a} ${value.b} ${value.d}',
           orElse: () => throw Error(),
         ),
         'a false 0.42',
       );
 
-      value = Complex.second('a', 21, 0.42);
+      value = SwitchTest.second('a', 21, 0.42);
 
       expect(
         value.maybeMap(
           null,
-          second: (Complex2 value) => '${value.a} ${value.c} ${value.d}',
+          second: (SwitchTest2 value) => '${value.a} ${value.c} ${value.d}',
           orElse: () => throw Error(),
         ),
         'a 21 0.42',
@@ -215,30 +216,30 @@ void main() {
     });
 
     test('assert orElse is passed', () {
-      var value = Complex('a');
+      var value = SwitchTest('a');
 
       expect(
         () => value.maybeMap((a) => '$a default', orElse: null),
         throwsA(isA<AssertionError>()),
       );
 
-      value = Complex.first('a', b: false, d: .42);
+      value = SwitchTest.first('a', b: false, d: .42);
 
       expect(
         () => value.maybeMap(
           null,
-          first: (Complex1 value) => '${value.a} ${value.b} ${value.d}',
+          first: (SwitchTest1 value) => '${value.a} ${value.b} ${value.d}',
           orElse: null,
         ),
         throwsA(isA<AssertionError>()),
       );
 
-      value = Complex.second('a', 21, 0.42);
+      value = SwitchTest.second('a', 21, 0.42);
 
       expect(
         () => value.maybeMap(
           null,
-          second: (Complex2 value) => '${value.a} ${value.c} ${value.d}',
+          second: (SwitchTest2 value) => '${value.a} ${value.c} ${value.d}',
           orElse: null,
         ),
         throwsA(isA<AssertionError>()),
@@ -246,15 +247,15 @@ void main() {
     });
 
     test('orElse is called', () {
-      var value = Complex('a');
+      var value = SwitchTest('a');
 
       expect(value.maybeMap(null, orElse: () => 42), 42);
 
-      value = Complex.first('a', b: false, d: .42);
+      value = SwitchTest.first('a', b: false, d: .42);
 
       expect(value.maybeMap(null, orElse: () => 42), 42);
 
-      value = Complex.second('a', 21, 0.42);
+      value = SwitchTest.second('a', 21, 0.42);
 
       expect(value.maybeMap(null, orElse: () => 42), 42);
     });
@@ -267,7 +268,7 @@ library main;
 import 'multiple_constructors.dart';
 
 void main() {
-  final value = Complex('a');
+  final value = SwitchTest('a');
 
   value.maybeMap(null, orElse: () => 42);
 }
@@ -290,7 +291,7 @@ library main;
 import 'multiple_constructors.dart';
 
 void main() {
-  final value = Complex('a');
+  final value = SwitchTest('a');
 
   value.maybeMap(null);
 }