+++ packages/freezed/CHANGELOG.md
@@ -12,6 +12,19 @@
   }
   ```
 
+- Added `Assert` decorator to generate `assert(...)` statements on Freezed classes:
+
+  ```dart
+  abstract class Person with _$Person {
+    @Assert('name.trim().isNotEmpty', 'name cannot be empty')
+    @Assert('age >= 0')
+    factory Person({
+      String name,
+      int age,
+    }) = _Person;
+  }
+  ```
+
 # 0.11.6
 
 - Fixed a bug where deep-copy did not work properly for recursive classes (https://github.com/rrousselGit/freezed/issues/213)